requirements.txt
py-tlsh>=4.7.2
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.20.9
